requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-crontab==0.7.1
django-request-logging==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
joblib==0.14.1
mongoengine==0.19.1
numpy==1.18.1
pandas==1.0.1
pycryptodomex==3.9.7
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
regex==2020.1.8
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
threadedprocess==0.0.5
tqdm==4.43.0
urllib3==1.25.8
xmltodict==0.12.0
threadedprocess~=0.0.5